+++ src/main/kotlin/mb/gitonium/GitoniumExtension.kt
@@ -1,6 +1,8 @@
 package mb.gitonium
 
 import org.gradle.api.Project
+import org.gradle.api.file.RegularFileProperty
+import org.gradle.internal.enterprise.test.FileProperty
 import java.util.regex.Pattern
 
 /** Extension for configuring the Gitonium plugin. */
@@ -49,6 +51,9 @@ open class GitoniumExtension(private val project: Project) {
 
     /** Whether to always create a snapshot version string, even if the HEAD points to a release tag. */
     var alwaysSnapshotVersion: Boolean = false
+
+    /** A properties file to write the build and version info to, or unset to not write. */
+    val buildPropertiesFile: RegularFileProperty = project.objects.fileProperty()
 
     /** The version info, determined lazily. */
     val versionInfo: GitoniumVersion by lazy {